LayoutTests/fast/frames/sandboxed-iframe-navigation-parent-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate a navigation change for frame with URL sandboxed-iframe-navigation-source.html.

This test verifies that a sandboxed IFrame cannot navigate an ancestor in the frame tree by assigning the location attribute.

This is done by loading ten non-sandboxed IFrames, and a single sandboxed one. Expect ten frames to be navigated, but the sandboxed one to not be one of them.

This test will print "PASS" on success.

               
PASS
